+++ BTD6-ai/Vector2.cpp
@@ -6,7 +6,7 @@ Vector2::Vector2()
 	Y = 0;
 }
 
-Vector2::Vector2(int newX, int newY)
+Vector2::Vector2(short newX, short newY)
 {
 	X = newX;
 	Y = newY;
